+++ QfP/LevelOne.pde
@@ -1,6 +1,6 @@
 int levelOne(){
   if(intro){
-    x = -1069/3;
+    x = (int)(-1069/3);
     
     String[] introText = loadStrings("intro.txt");
     background(0);
